+++ SEDAPExpress/src/main/java/de/bundeswehr/mese/sedapexpress/messages/OWNUNIT.java
@@ -129,6 +129,7 @@ public class OWNUNIT extends SEDAPExpressMessage {
     }
 
     /**
+     * Instantiate a new OWNUNIT message
      *
      * @param number
      * @param time
@@ -147,7 +148,7 @@ public class OWNUNIT extends SEDAPExpressMessage {
      * @param name
      * @param sidc
      */
-    public OWNUNIT(Byte number, Long time, String sender, Character classification, Boolean acknowledgement, String mac,
+    public OWNUNIT(Short number, Long time, String sender, Character classification, Boolean acknowledgement, String mac,
 	    Double latitude, Double longitude, Double altitude, Double speed, Double course, Double heading,
 	    Double roll, Double pitch, String name, char[] sidc) {
 
@@ -166,6 +167,7 @@ public class OWNUNIT extends SEDAPExpressMessage {
     }
 
     /**
+     * Instantiate a new OWNUNIT message from a serialized message
      *
      * @param message
      */
@@ -175,6 +177,7 @@ public class OWNUNIT extends SEDAPExpressMessage {
     }
 
     /**
+     * Instantiate a new OWNUNIT message from a paramter list
      *
      * @param message
      */
